+++ src/app/watch.py
@@ -6,8 +6,7 @@ import typer
 from loguru import logger
 from watchfiles import run_process
 
-from app.constants import ColorScheme
-from app.build import build, build_page, __file__ as build_module_file, get_html_path
+from app.build import build_page, __file__ as build_module_file, get_html_path
 
 
 cli = typer.Typer()
@@ -15,7 +14,6 @@ cli = typer.Typer()
 
 @cli.callback(invoke_without_command=True)
 def page(
-    color: Annotated[ColorScheme, typer.Option(help="Render with color scheme.")] = ColorScheme.light,
     auto_refresh: Annotated[bool, typer.Option(help="Force a refresh every 3 seconds.")] = False,
 ):
     html_path = get_html_path()
@@ -25,10 +23,10 @@ def page(
     watch_paths = [build_module_file, html_path, static_path, readme_path]
 
     logger.debug("Triggering initial build")
-    build_page(color, debug=True)
+    build_page(debug=True)
 
     logger.debug("Opening browser to page")
     subprocess.run(["explorer.exe", str(html_path)])
 
     logger.debug(f"Watching for changes in {watch_paths}")
-    run_process(*watch_paths, target=build_page, args=(color,), kwargs=dict(debug=auto_refresh))
+    run_process(*watch_paths, target=build_page, kwargs=dict(debug=auto_refresh))