+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## v5.3.0 - 2025-03-14
+* Added additional sizes (medium, large)
+* Set medium size to default
+* Added size selector
+* Moved download button
+
 ## v5.2.1 - 2025-03-09
 * Fixed typo on resume